bump pydantic
